adding driver,admin controller

+++ next-server/src/main/java/com/noteam/next/services/DriverService.java
@@ -1,7 +1,13 @@
 package com.noteam.next.services;
 import com.noteam.next.entities.Driver;
+import com.noteam.next.entities.Order;
+import com.noteam.next.entities.State;
 import com.noteam.next.repositories.DriverRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -27,13 +33,24 @@ public class DriverService {
         return driverRepository.findByEmail(email);
     }
 
+    public List<Driver> getAllDrivers(String sortBy, String sortDir){
+        logger.info("Driver service: getting all Drivers sorted by "+ sortBy+",("+ sortDir+")");
+        return driverRepository.findAll(Sort.by(Sort.Direction.fromString(sortDir),sortBy));
+    }
+
+    public Page<Driver> getDdriversByPage(int page, int size, String sortBy, String sortDir){
+        Pageable pageable = PageRequest.of(page,size,Sort.by(Sort.Direction.fromString(sortDir),sortBy));
+        logger.info("Driver service: getting Drivers by page: "+page+" with size "+size+" sorted by "+ sortBy+",("+ sortDir+")");
+            return driverRepository.findAll(pageable);
+    }
+
     public List<Driver> findAll() {
         logger.info("Getting all drivers");
         return driverRepository.findAll();
     }
 
     // post
-    public int createdriver(String name, int age ,String image, String social_security_number, String email, String password, Boolean isbusy){
+    public Driver createdriver(String name, int age ,String image, String social_security_number, String email, String password, Boolean isbusy){
         logger.info("creating a new driver");
         Driver driver = new Driver();
         driver.setName(name);
@@ -45,15 +62,14 @@ public class DriverService {
         driver.setIsbusy(false);
         driver.setCreated_at(LocalDateTime.now());
         driver.setUpdated_at(LocalDateTime.now());
-        driverRepository.save(driver);
-        return 1;
+        return driverRepository.save(driver);
     }
 
     public int updatedriver (int driver_id, String name, int age , String image, String social_security_number, String email, String password, Boolean isbusy){
         Optional<Driver> driverOptional= getDriverById(driver_id);
         if(driverOptional.isEmpty()) {
             logger.info("Driver service: The driver with id: " + driver_id + " is not found!");
-            return -1;
+            return 0;
         }
         else{
                 logger.info("updating driver by id : " + driver_id);
@@ -71,8 +87,15 @@ public class DriverService {
             }
         }
     // delete
-    public void deleteById(int driver_id) {
-        logger.info("Deleting the Driver by id: " + driver_id );
-        driverRepository.deleteById(driver_id);
+    public boolean deleteDriverById(int driver_id){
+        Optional<Driver> driverOptional = getDriverById(driver_id);
+        if (driverOptional.isEmpty()){
+            return false;
+        }
+        else {
+            driverRepository.deleteById(driver_id);
+            return true;
+        }
     }
+
 }
